Fix inlined kind param
It seems as though the kind param isn't making it through our JSON parser. My suspicion is that it's getting inlined by the Kotlin compiled somehow since switching the kind to be retrieved via a function fixed the problem. Note: The function had to NOT be a getter/setter since it seemed to suffer from the same problem